Feature/transactions (Depends on PR #37)

+++ Sources/CoreDataRepository/CoreDataRepository+Create.swift
@@ -13,22 +13,26 @@ extension CoreDataRepository {
         _ item: Model,
         transactionAuthor: String? = nil
     ) async -> Result<Model, CoreDataError> where Model: WritableUnmanagedModel, Model: FetchableUnmanagedModel {
-        await context.performInScratchPad(schedule: .enqueued) { [context] scratchPad in
+        let context = Transaction.current?.context ?? context
+        let notTransaction = Transaction.current == nil
+        return await context.performInScratchPad(schedule: .enqueued) { [context] scratchPad in
             scratchPad.transactionAuthor = transactionAuthor
             let object = try item.asManagedModel(in: scratchPad)
             let tempObjectId = object.objectID
             try item.updating(managed: object)
             try scratchPad.save()
-            try context.performAndWait {
-                context.transactionAuthor = transactionAuthor
-                do {
-                    try context.save()
-                } catch {
-                    let parentContextObject = context.object(with: tempObjectId)
-                    context.delete(parentContextObject)
-                    throw error
+            if notTransaction {
+                try context.performAndWait {
+                    context.transactionAuthor = transactionAuthor
+                    do {
+                        try context.save()
+                    } catch {
+                        let parentContextObject = context.object(with: tempObjectId)
+                        context.delete(parentContextObject)
+                        throw error
+                    }
+                    context.transactionAuthor = nil
                 }
-                context.transactionAuthor = nil
             }
             try scratchPad.obtainPermanentIDs(for: [object])
             return try Model(managed: object)
